Copy wymplugin.js to etherpadplugin.js
Now integrated with CollabClient in a meaningful way.  The client
can connect to the server and receive push messages much like we
could do with WebSocket.  Messages are parsed and logged in a more
meaningful way to our editor.  Unable to push messages to the server
as of now, something is causing collabclient to not actually send
messages, even though they are being queued.

Signed-off-by: Kevin Wells <wellska1@rose-hulman.edu>